requirements.txt
fastapi
fastapi[standard]
pandas
uvicorn
asyncpg
dotenv
tortoise-orm[asyncpg]
passlib
aioredis
